requirements.txt
flask
babel
python-dateutil
flask-moment
flask-wtf
flask-sqlalchemy
flask-migrate
psycopg2-binary
sqlalchemy